Separated out the definitions to clean up a bit
We're going to be using C to compile our Lisp eventually, so I need to
have a header file for compilation to work.  We can still separate out
the implementation details into separate files (then generate a shared
library for usage in the link stage) but we need this header file to
be singular and clean.